add base class for keras model and have sample model inherit from that
